[sonamu] SON-444 audit-log-plugin 모듈 분리 및 빌더 중복 제거
